data-lambda-functions/7001d_Lambda/requirements.txt
-i https://pypi.org/simple
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.0.8
backcall==0.2.0
bcrypt==4.0.0; python_version >= '3.6'
boto3==1.24.66
botocore==1.27.66; python_version >= '3.7'
cffi==1.15.1
cryptography==37.0.4; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.5'
executing==1.0.0
ipython==8.4.0
patool == 1.12
jedi==0.18.1; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
numpy==1.23.2; python_version >= '3.10'
pandas==1.4.4
paramiko==2.11.0
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.13.0; python_version >= '3.6'
pynacl==1.5.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2.1
rarfile==4.0
s3transfer==0.6.0; python_version >= '3.7'
setuptools==65.3.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.5.0
traitlets==5.3.0; python_version >= '3.7'
unrar==0.4
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
wcwidth==0.2.5
